Polynomial calibration
The previous Arduino sketch had an adjustment for the an observed effect where the readings are too low at high irradiances and too high at low irradiances. This was flawed. The code now supports a better way to compensate for this effect which is to calibrate based on a polynomial (quadratic) curve rather than a straight line. The PYRANO_CAL_A constant is the x^2 coefficient. If A=0 (the default), the calibration is linear.

The document was updated to reflect this change.